+++ common/views/components/MediaObject/README.md
@@ -1,0 +1,5 @@
+## Purpose
+
+Media object to display icons and text in a grid layout format.
+
+[Edit this on GitHub](https://github.com/wellcomecollection/wellcomecollection.org/edit/master/common/views/components/MediaObject/README.md)
