+++ common/views/components/MediaObjectBase/README.md
@@ -1,0 +1,5 @@
+## Purpose
+
+The base component which use currently being used in MediaObject, CompactCard
+
+[Edit this on GitHub](https://github.com/wellcomecollection/wellcomecollection.org/edit/master/common/views/components/MediaObjectBase/README.md)
